sc-10744 reorganization of the admin UI project
