+++ web/gds-admin-ui/src/pages/app/details/BasicDetails/components/__test__/trisa-details.spec.js
