+++ web/gds-admin-ui/src/pages/app/details/__tests__/basic-details.spec.js
