+++ web/gds-admin-ui/src/utils/__test__/utils.spec.js
@@ -1,4 +1,4 @@
-import { defaultEndpointPrefix } from "../utils";
+import { defaultEndpointPrefix } from "..";
 
 describe("defaultEndpointPrefix", () => {
